# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: engineer
        type: string
    type: object
...
  models.Equipment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      installation_date:
        type: string
      inventory_number:
        type: string
      last_maintenance_date:
        type: string
      location:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      serial_number:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  services.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API веб-приложения для управления техническим обслуживанием и ремонтом
    оборудования предприятия
  title: ТОиР API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Получение JWT токенов по логину и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.TokenPair'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Аутентификация пользователя
      tags:
      - auth
...
  /equipment:
    get:
      description: Получение списка оборудования с пагинацией и фильтрами
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Элементов на странице
        in: query
        name: per_page
        type: integer
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Фильтр по расположению
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Equipment'
                  type: array
                meta:
                  $ref: '#/definitions/response.Meta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Список оборудования
      tags:
      - equipment
    post:
      consumes:
      - application/json
      description: Создаёт новую единицу оборудования
      parameters:
      - description: Данные оборудования
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Equipment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Equipment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание оборудования
      tags:
      - equipment
  /equipment/{id}:
    delete:
      description: Удаляет единицу оборудования по ID
      parameters:
      - description: ID оборудования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Удаление оборудования
      tags:
      - equipment
    get:
      description: Возвращает оборудование по идентификатору
      parameters:
      - description: ID оборудования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Equipment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение оборудования по ID
      tags:
      - equipment
    put:
      consumes:
      - application/json
      description: Обновляет данные единицы оборудования
      parameters:
      - description: ID оборудования
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Equipment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Equipment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновление оборудования
      tags:
      - equipment
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"